backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
torchvision
numpy
opencv-python
